+++ src/main/java/org/chiu/megalith/search/mq/handler/RemoveBlogIndexHandler.java
@@ -83,8 +83,8 @@ public final class RemoveBlogIndexHandler extends BlogIndexSupport {
         //设置getBlogDetail的bloom
         redisTemplate.opsForValue().setBit(BLOOM_FILTER_BLOG.getInfo(), blog.getId(), false);
         //重置该年份的页面bloom
-        Integer countByPeriod = blogRepository.countByCreatedBetween(start, end);
-        int totalPageByPeriod = countByPeriod % blogPageSize == 0 ? countByPeriod / blogPageSize : countByPeriod / blogPageSize + 1;
+        Long countByPeriod = blogRepository.countByCreatedBetween(start, end);
+        int totalPageByPeriod = (int) (countByPeriod % blogPageSize == 0 ? countByPeriod / blogPageSize : countByPeriod / blogPageSize + 1);
         for (int i = 1; i <= totalPageByPeriod; i++) {
             redisTemplate.opsForValue().setBit(BLOOM_FILTER_YEAR_PAGE.getInfo() + year, i, true);
         }
